feat(attributes): add ability to copy attributes between categories
This commit introduces a new feature allowing administrators to copy all attributes from a source category to a destination category.

A new "Copy Attributes" button and modal have been added to the admin UI. This allows for the selection of a source and destination category to perform the copy operation.

The backend is supported by a new API endpoint `POST /api/admin/attributes/copy` which handles the duplication of attribute records in the database. This streamlines the process of setting up new categories that share similar attributes.